게시글 컨트롤러에 jwt 적용 및 예외 처리 추가

+++ src/main/java/com/example/basicboard/service/PostService.java
@@ -1,9 +1,12 @@
 package com.example.basicboard.service;
 
 import com.example.basicboard.advice.exception.CPostNotFoundException;
+import com.example.basicboard.advice.exception.CUserNotFoundException;
 import com.example.basicboard.dto.PostRequestDto;
 import com.example.basicboard.models.Post;
+import com.example.basicboard.models.User;
 import com.example.basicboard.repository.PostRepository;
+import com.example.basicboard.repository.UserRepository;
 import com.sun.istack.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,35 +17,59 @@ import java.util.*;
 public class PostService {
 
     private final PostRepository postRepository;
+    private final UserRepository userRepository;
 
     @Autowired
-    public PostService(PostRepository postRepository){
+    public PostService(PostRepository postRepository, UserRepository userRepository){
         this.postRepository = postRepository;
+        this.userRepository = userRepository;
     }
 
     public List<Post> getAllPost() {
         return postRepository.findAll();
     }
 
-    public Post getPost(@NotNull long postId) {
-        return postRepository.findById(postId).orElseThrow(CPostNotFoundException::new);
+    public List<Post> getUserPosts(String userId) {
+        User user = userRepository.findByUserId(userId).orElseThrow(CUserNotFoundException::new);
+        return postRepository.findAllByUser(user);
     }
 
-    public Long createPost(@NotNull PostRequestDto postRequestDto){
-        Post post = new Post(postRequestDto);
+    public Post getPost(@NotNull long postId, String userId) {
+        Post post = postRepository.findById(postId).orElseThrow(CPostNotFoundException::new);
+        User writer = post.getUser();
+        User user = userRepository.findByUserId(userId).orElseThrow(CUserNotFoundException::new);
+        if(!writer.equals(user)){
+            throw new CPostNotFoundException();
+        }
+        return post;
+    }
+
+    public Long createPost(@NotNull PostRequestDto postRequestDto, String userId){
+        User user = userRepository.findByUserId(userId).orElseThrow(CUserNotFoundException::new);
+        Post post = new Post(postRequestDto, user);
         postRepository.save(post);
         return post.getId();
     }
 
-    public Long updatePost(@NotNull long postId, @NotNull PostRequestDto postRequestDto){
+    public Long updatePost(@NotNull long postId, @NotNull PostRequestDto postRequestDto, String userId){
         Post post = postRepository.findById(postId).orElseThrow(CPostNotFoundException::new);
+        User user = userRepository.findByUserId(userId).orElseThrow(CUserNotFoundException::new);
+        User writer = post.getUser();
+        if(!writer.equals(user)){
+            throw new CPostNotFoundException();
+        }
         post.update(postRequestDto);
         postRepository.save(post);
         return post.getId();
     }
 
-    public long deletePost(long postId) {
+    public long deletePost(long postId, String userId) {
         Post post = postRepository.findById(postId).orElseThrow(CPostNotFoundException::new);
+        User writer = post.getUser();
+        User user = userRepository.findByUserId(userId).orElseThrow(CUserNotFoundException::new);
+        if(!writer.equals(user)){
+            throw new CPostNotFoundException();
+        }
         postRepository.delete(post);
         return postId;
     }
